+++ app/controllers/hq/partners_controller.rb
@@ -1,6 +1,4 @@
-class Hq::PartnersController < ApplicationController
-  before_action :authenticate_admin_user!
-  layout 'application'
+class Hq::PartnersController < HqController
 
   def index
     @partners= Partner.all
