+++ app/controllers/hq_controller.rb
@@ -1,0 +1,5 @@
+class HqController < ApplicationController
+  before_action :authenticate_admin_user!
+  layout 'application'
+end
+
